requirements_dev.txt
coverage~=6.1
django-stubs~=1.9.0
lorem~=0.1.1
mypy~=0.910
pre-commit~=2.15.0
pylint~=2.11.1
pylint-django~=2.4.4
types-paramiko~=2.7.4
